# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Next Olshop API",
    "description": "REST API for Next Olshop — Single Store Online Shop untuk UMKM. Digunakan oleh Flutter mobile app.\n\n## Authentication\nSemua endpoint yang membutuhkan auth menggunakan Bearer token dari `/api/auth/login`.\n\n## Response Format\n- **Success**: `{ success: true, data: T }`\n- **Error**: `{ success: false, error: string }`\n- **Pagination**: `{ success: true, data: T[], meta: { page, limit, total, totalPages } }`\n\n## Catatan\n- Harga dalam Rupiah (IDR), disimpan sebagai string decimal (`\"299000.00\"`)\n- Tanggal dalam format ISO 8601 (UTC)\n- Berat produk dalam gram",
    "version": "1.0.0",
    "contact": {
      "name": "Next Olshop",
      "url": "https://github.com/codewithbahri"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token dari POST /api/auth/login"
      }
    },
    "schemas": {
...
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "userId": { "type": "integer", "example": 2 },
          "orderNumber": { "type": "string", "example": "ORD2604ABC1" },
          "status": { "type": "string", "enum": ["waiting_payment", "packing", "shipping", "delivered", "expired", "cancelled"], "example": "waiting_payment" },
          "subtotal": { "type": "string", "example": "598000.00" },
          "shippingCost": { "type": "string", "example": "45000.00" },
          "total": { "type": "string", "example": "643000.00" },
          "shippingName": { "type": "string", "example": "John Doe" },
          "shippingPhone": { "type": "string", "example": "08123456789" },
          "shippingAddress": { "type": "string", "example": "Jl. Raya Sudirman No. 123" },
          "notes": { "type": "string", "nullable": true },
          "willExpiredAt": { "type": "string", "format": "date-time", "nullable": true },
          "paidAt": { "type": "string", "format": "date-time", "nullable": true },
          "shippingAt": { "type": "string", "format": "date-time", "nullable": true },
          "deliveredAt": { "type": "string", "format": "date-time", "nullable": true },
          "expiredAt": { "type": "string", "format": "date-time", "nullable": true },
          "cancelledAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrderItem" }
          },
          "invoice": { "$ref": "#/components/schemas/Invoice" },
          "shipping": { "$ref": "#/components/schemas/Shipping" }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "orderId": { "type": "integer" },
          "productId": { "type": "integer" },
          "productName": { "type": "string", "example": "Wireless Earbuds Pro" },
          "productImage": { "type": "string", "nullable": true },
          "price": { "type": "string", "example": "299000.00" },
          "quantity": { "type": "integer", "example": 2 },
          "subtotal": { "type": "string", "example": "598000.00" }
        }
      },
      "Invoice": {
        "type": "object",
        "nullable": true,
        "properties": {
          "id": { "type": "integer" },
          "orderId": { "type": "integer" },
          "xenditId": { "type": "string", "example": "inv_123456" },
          "invoiceUrl": { "type": "string", "example": "https://checkout.xendit.co/invoices/inv_123456" },
          "paymentMethod": { "type": "string", "nullable": true },
          "paymentChannel": { "type": "string", "nullable": true },
          "amount": { "type": "string", "example": "643000.00" },
          "status": { "type": "string", "enum": ["pending", "paid", "expired", "cancelled"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Shipping": {
        "type": "object",
        "nullable": true,
        "properties": {
          "id": { "type": "integer" },
          "orderId": { "type": "integer" },
          "recipientName": { "type": "string" },
          "phone": { "type": "string" },
          "address": { "type": "string" },
          "courierName": { "type": "string", "example": "J&T Regular" },
          "courierCompany": { "type": "string", "example": "jnt" },
          "courierType": { "type": "string", "example": "REG" },
          "price": { "type": "string", "example": "45000.00" },
          "trackingId": { "type": "string", "nullable": true, "example": "TRK123456" },
          "waybillId": { "type": "string", "nullable": true },
          "estimateDays": { "type": "string", "nullable": true, "example": "3-4" },
          "status": { "type": "string", "nullable": true, "example": "pending_pickup" }
        }
      },
      "ShippingHistory": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "shippingId": { "type": "integer" },
          "status": { "type": "string" },
          "note": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Error message" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Order Detail",
        "description": "Ambil detail pesanan lengkap dengan items, invoice, shipping, dan tracking histories",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "Order ID" }
        ],
        "responses": {
          "200": {
            "description": "Detail order",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "allOf": [
                        { "$ref": "#/components/schemas/Order" },
                        {
                          "type": "object",
                          "properties": {
                            "trackingHistories": {
                              "type": "array",
                              "items": { "$ref": "#/components/schemas/ShippingHistory" }
                            }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Bukan pesanan milik user",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Pesanan tidak ditemukan",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/orders/{id}/cancel": {
      "post": {
        "tags": ["Orders"],
        "summary": "Cancel Order",
        "description": "Batalkan pesanan. Hanya bisa untuk status waiting_payment atau packing. Stock akan dikembalikan.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "Order ID" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": { "type": "string", "example": "Berubah pikiran", "description": "Alasan pembatalan (opsional)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order dibatalkan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "status": { "type": "string", "example": "cancelled" },
                        "cancelledAt": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Order tidak bisa dibatalkan dari status ini",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Bukan pesanan milik user",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Pesanan tidak ditemukan",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/orders/{id}/repay": {
      "post": {
        "tags": ["Orders"],
        "summary": "Repay Expired Order",
        "description": "Bayar ulang order yang sudah expired. Membuat invoice Xendit baru dan mengembalikan paymentUrl.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "Order ID" }
        ],
        "responses": {
          "200": {
            "description": "Invoice baru dibuat, redirect ke paymentUrl",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "paymentUrl": { "type": "string", "example": "https://checkout.xendit.co/invoices/inv_789012" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Order tidak bisa dibayar ulang dari status ini",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Bukan pesanan milik user",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Pesanan tidak ditemukan",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/orders/{id}/tracking": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get Order Tracking",
        "description": "Ambil informasi tracking pengiriman pesanan",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "Order ID" }
        ],
        "responses": {
          "200": {
            "description": "Tracking info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "trackingId": { "type": "string", "example": "TRK123456" },
                        "status": { "type": "string", "example": "in_transit" },
                        "histories": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/ShippingHistory" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Tracking tidak ditemukan",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
  }
}